requirements.txt
streamlit==1.39.0
plotly==5.10.0
numpy==1.23.5
bcrypt==4.0.1
openpyxl==3.1.2
firebase-admin==6.2.0
google-auth>=2.22.0
google-cloud-firestore>=2.11.1
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dateutil>=2.8.2
pytz>=2023.3
